Reworked the Kafka Topic Structure.
Rather than using Topics that include the job id the name, we will instead use more generic topics.  We will then include the jobId inside the message being sent to the topic.

This enables us to create more generic consumers.  The consumers don't need to know in advance what the jobs will be.